refactor(proxy-gateway): split OpenAI request and response conversion out of the service
Group 3 of the gateway port. Behavior preserving, no API change.

`openai.service.ts` was the largest file in the gateway at 1394 lines, and about
400 of them were pure mapping between the OpenAI surface and the Claude shape
the gateway speaks internally: request conversion, tool conversion, finish
reason mapping, tool argument normalisation and the response conversion back.

Those ten functions never touched instance state. They referenced only each
other, no logger, no injected service, which is what made this a mechanical move
rather than a redesign. They now live in
`modules/openai/chat/openai-claude-conversion.ts` as free functions, and the
service delegates.

The protocol boundary is unchanged on purpose. `OpenAIService` stays the owner
of the OpenAI surface and keeps the orchestration: retry loop, account lease,
stream construction, upstream calls. Only the mapping moved, and it moved into
the OpenAI module rather than into anything shared, because it is
OpenAI-specific by definition.

Three thin private delegates remain on the service: `convertOpenAIToClaude`,
`convertClaudeToOpenAIResponse` and `convertOpenAIToolsToAnthropicTools`. The
parity, retry and tool-mapper suites reach them through the service instance,
two of them by `Reflect.get`, and a split meant to preserve behavior should
preserve the reachable surface as well. Removing them is a separate decision
about those tests, not something to do silently here.

  openai.service.ts               1394 -> 1007 lines
  chat/openai-claude-conversion.ts    new, 440 lines

Verified: `tsc --noEmit` clean, `eslint .` 0 errors with the same 8 warnings as
the parent, `prettier --check .` clean, `vitest run` 815 passed and 4 skipped
with no failures.

A note on measuring this locally: the full suite has to be run with
`--maxWorkers=4` on this machine. At the default worker count it produces
timeout failures that are contention, not signal. The same run on the parent
commit produced 21 such failures with zero assertion failures, which is how they
were told apart.
